impl navigation and playback

+++ app/src/main/java/com/example/flow/data/remote/response_models/GetNextSongResponseModels.kt
@@ -1,11 +1,14 @@
 package com.example.flow.data.remote.response_models
 
 // TODO add JSON mappers, android field names might differ from API field names
+/**
+ * API response for each song.
+ */
 data class SongWithUrl(
     val id: Int,
     val title: String,
     val artistStr: String,
-    val durationMillis: Long,
+    val durationMillis: Int,
     val albumArtUrl: String,
     val songUrl: String
 )
